MA_Logstash/auth_service/requirements.txt
flask
pika
requests
python-json-logger
PyJWT